- Refactored state updates in ProduitForm to use functional updates for better performance.
- Added console logging for file selection in ProduitForm.
- Improved error handling in ProduitList to provide more informative feedback to users.
- Updated form submission logic in ProduitList to ensure all required fields are validated before submission.
- Introduced RecetteForm and RecetteList components for managing recipes, including form handling and data fetching.
- Enhanced UI components with better styling and layout adjustments.
- Fixed image source fallback in FamilleTypeCarousels.
- Improved table rendering in TableMui with better styling and structure.

+++ BackEnd/database/migrations/2024_12_12_102024_prix_produit.php
@@ -11,17 +11,17 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('prix_produits', function (Blueprint $table) {
-            $table->id(); // ID auto-incrémenté pour la table
-            $table->unsignedBigInteger('produit_id')->nullable(); // Référence au produit
-            $table->date('dateDebut')->nullable(); // Date de début
-            $table->date('dateFin')->nullable(); // Date de fin
-            $table->decimal('prixProduit', 10, 2)->nullable(); // Prix du produit avec deux décimales
-            $table->timestamps(); // Colonnes created_at et updated_at
-
-            // Clé étrangère vers la table des produits (si elle existe)
-            $table->foreign('produit_id')->references('id')->on('produits')->onDelete('cascade');
-        });
+        if (!Schema::hasTable('prix_produits')) {
+            Schema::create('prix_produits', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('produit_id')->nullable();
+                $table->date('dateDebut')->nullable();
+                $table->date('dateFin')->nullable();
+                $table->decimal('prixProduit', 10, 2)->nullable();
+                $table->timestamps();
+                $table->foreign('produit_id')->references('id')->on('produits')->onDelete('cascade');
+            });
+        }
     }
 
     /**
